stamp-extractor-backend/requirements.txt
fastapi
uvicorn
pymupdf
pillow
opencv-python-headless
numpy
pytesseract
python-multipart
